Added server synchronization between clients
Added backend socket event that emits the current members and chat messages to the other members in the room, currently runs on member join and message sending. Upon page refresh, roomToken is taken from session storage and given to the server. Server then returns the user information and the state variables are set to the returned info.